+++ game/CrystalBall.cpp
@@ -69,7 +69,7 @@ void CrystalBall::Render(Camera * camera)
 
 	D3DXVECTOR2 pos = camera->Transform(x, y);
 
-	_sprite->Draw((int)pos.x, (int)pos.y);
+	_sprite->Draw(pos.x, pos.y);
 
 	if (IS_DEBUG_RENDER_BBOX)
 		RenderBoundingBox(camera);
