+++ game/Dagger.cpp
@@ -42,7 +42,7 @@ void Dagger::Create(float simonX, float simonY, int simonDirection)
 }
  
 
-void Dagger::RenderIcon(int X, int Y)
+void Dagger::RenderIcon(float X, float Y)
 {
 	_sprite->DrawFlipXIndex(0, X, Y); 
 }
