+++ game/MorningStar.cpp
@@ -43,9 +43,9 @@ void MorningStar::Render(Camera * camera)
 {
 	D3DXVECTOR2 pos = camera->Transform(x, y);
 	if (direction == -1)
-		_sprite->Draw((int)pos.x, (int)pos.y);
+		_sprite->Draw(pos.x, pos.y);
 	else
-		_sprite->DrawFlipX((int)pos.x, (int)pos.y);
+		_sprite->DrawFlipX(pos.x, pos.y);
 
 
 	if (IS_DEBUG_RENDER_BBOX)
@@ -175,7 +175,7 @@ void MorningStar::GetBoundingBox(float & left, float & top, float & right, float
  
 }
 
-void MorningStar::RenderIcon(int X, int Y)
+void MorningStar::RenderIcon(float X, float Y)
 {
 }
